fix: move lifecycle types to model, validate days param, clarify COALESCE (#37)
- Move LifecycleSummary and LifecycleSnapshot structs from store to model
- Return 400 for non-numeric days query param, keep range clamping
- Add comment explaining COALESCE behavior for zero-flag projects
- Initialize empty slice in GetTrends to avoid nil check in handler
